+++ austrakka/components/field/funcs.py
@@ -2,10 +2,11 @@ import pandas as pd
 
 from loguru import logger
 
-from austrakka.utils.helpers.fields import get_fieldtype_by_name, get_field_by_name
-
-from austrakka.utils.api import call_api
-from austrakka.utils.api import get, post, put
+from austrakka.utils.helpers.fieldtype import get_fieldtype_by_name
+from austrakka.utils.helpers.fields import get_field_by_name
+from austrakka.utils.api import api_get
+from austrakka.utils.api import api_post
+from austrakka.utils.api import api_put
 from austrakka.utils.misc import logger_wraps
 from austrakka.utils.output import print_table
 from austrakka.utils.paths import METADATACOLUMN_PATH
@@ -16,8 +17,7 @@ def list_fields(out_format: str):
     """
     List all metadata fields (MetaDataColumns) within AusTrakka.
     """
-    response = call_api(
-        method=get,
+    response = api_get(
         path=METADATACOLUMN_PATH,
     )
 
@@ -71,10 +71,9 @@ def add_field(
                 f"set --colour-nodes.")
         can_visualise = (typename not in ["date", "number", "string"])
 
-    call_api(
-        method=post,
+    api_post(
         path=METADATACOLUMN_PATH,
-        body={
+        data={
             "ColumnName": name,
             "CanVisualise": can_visualise,
             "ColumnOrder": column_order,
@@ -129,19 +128,18 @@ def update_field(
                     f"Setting colour-nodes flag on field {name} of type {typename}. "
                     f"This may work poorly as colour visualisations are configured for a "
                     f"small discrete set of values.")
-        post_field["canVisualise"] = (can_visualise == 'viz')
+        post_field["canVisualise"] = can_visualise == 'viz'
 
     if column_order is not None:
         post_field["columnOrder"] = column_order
 
     if set_show is not None:
-        post_field["isDisplayedAsDefault"] = (set_show == 'show')
+        post_field["isDisplayedAsDefault"] = set_show == 'show'
 
     if min_width is not None:
         post_field["minWidth"] = min_width
 
-    call_api(
-        method=put,
+    api_put(
         path=f"{METADATACOLUMN_PATH}/{field['metaDataColumnId']}",
-        body=post_field
+        data=post_field
     )