+++ austrakka/utils/helpers/output.py
@@ -1,15 +1,18 @@
+from typing import Dict
+
 import pandas as pd
 
-from austrakka.utils.api import call_api, get
+from austrakka.utils.api import api_get
 from austrakka.utils.misc import logger_wraps
 from austrakka.utils.output import print_table
 
 
 @logger_wraps()
-def call_get_and_print_table(path: str, out_format: str):
-    response = call_api(
-        method=get,
+def call_get_and_print_table(path: str, out_format: str, params: Dict = None):
+    params = {} if params is None else params
+    response = api_get(
         path=path,
+        params=params,
     )
 
     result = response['data'] if ('data' in response) else response